+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## Unreleased
 - Added - cli.success() and cli.fail() overload with just the exit code
+- Added - cli.parse() overload taking a const string vector
 
 ## dimcli 7.4.0 (2025-06-16)
 - Added - cli.beforeExec() and cli.afterExec() for code common to all commands